fix goreleaser/release issues
