+++ controller.Dockerfile
@@ -2,7 +2,7 @@
 # Refer to https://github.com/GoogleContainerTools/distroless for more details
 FROM gcr.io/distroless/static:nonroot
 WORKDIR /
-COPY bin/manager .
+COPY manager .
 USER 65532:65532
 
 ENTRYPOINT ["/manager"]
